Fix Join::into_ref implementation and report failures
It forgot to clear the given buffer when it was consumed but the joined
buffer was not yet available. This affect with the default configuration
because the last iteration (the one checked) is a joined buffer if the
number of iterations (i.e. 10^7) is dividable by the slice count (i.e.
2^4).

To check the result of a non-adjoined last buffer run e.g.:

    ITERATIONS=1000 RECORD_SIZE=1024 SLICE_COUNT=16 ./build_and_run.sh