Reduce the record size indexing overhead
By enforcing the size of record size patterns to be specified in
power-of-2s, we can optimize the indexing to use binary masking instead
of modulus. This is the sort of optimization that compilers do
automatically for statically known power-of-2 sizes, but here the size
is a given parameter so they cannot optimize it themselves.